refactor: move db related item in separate folder
Author: Rudy Savini - @SavGRY
