+++ backend/alembic/env.py
@@ -27,7 +27,7 @@ if config.config_file_name is not None:
 # for 'autogenerate' support
 # from myapp import mymodel
 # target_metadata = mymodel.Base.metadata
-from core.database import Base
+from core.db.database import Base
 
 target_metadata = Base.metadata
